+++ common/src/main/kotlin/me/znotchill/marmot/common/ui/components/SpriteComponent.kt
@@ -13,15 +13,15 @@ open class SpriteComponent(
     override val compType: CompType
         get() = CompType.SPRITE
 
-    override fun width(): Int {
+    override fun width(): Float {
         val drawWidth = if (props.size.x > 0) props.size.x else (computedSize?.x ?: 0f)
         val baseWidth = drawWidth + props.padding.left + props.padding.right
-        return (baseWidth * props.scale.x).toInt()
+        return (baseWidth * props.scale.x)
     }
 
-    override fun height(): Int {
+    override fun height(): Float {
         val drawHeight = if (props.size.y > 0) props.size.y else (computedSize?.y ?: 0f)
         val baseHeight = drawHeight + props.padding.top + props.padding.bottom
-        return (baseHeight * props.scale.y).toInt()
+        return (baseHeight * props.scale.y)
     }
 }
